Fixes tests, adds null checks

+++ src/Umbraco.Tests/UmbracoExamine/IndexInitializer.cs
@@ -74,9 +74,10 @@ namespace Umbraco.Tests.UmbracoExamine
 
                 mediaService = Mock.Of<IMediaService>(
                     x => x.GetPagedDescendants(
-                        It.IsAny<int>(), It.IsAny<int>(), It.IsAny<int>(), out totalRecs, It.IsAny<string>(), It.IsAny<Direction>(), It.IsAny<bool>(), It.IsAny<string>()) 
+                        It.IsAny<int>(), It.IsAny<long>(), It.IsAny<int>(), out totalRecs, It.IsAny<string>(), It.IsAny<Direction>(), It.IsAny<bool>(), It.IsAny<string>()) 
                         ==
                         allRecs);
+                
             }
             if (dataTypeService == null)
             {
